Match Making Update
I built more of the UI for the Match making elements.
I added to the SceneController so that the game can be quit from pressing a button in the main menu.
Added a function in the PhotonConnector to try to join a random room and if it doesn't work then it creates one.
Added the ability to create a custom room, still open currently, for getting friends to join.
Set up a button to allow an AI match.
Created two extra scenes to house the two different games, either with another person or with the AI, may need only one.
Created a room photon manager and room leave/quite button prefab.