Match Making Update
I Updated the UIFriend to listen for a room status update allowing the invite button to show up when the player is in a room and their friend is online. And to check whether the inviteButton is active prior to setting it active. (May need to look over again.)
I updated the  PhotonConnector by removing any unused code to simplify what I am working with.
I created a UIRoomInvite script that listens for an acceptance of a room invite and then adjusts the UI accordingly to show it.