Match Making Update
I made it so that you can enter a Lobby that waits for a friend. When the friend joins, you can enter the game with them.
You can also enter a random open lobby or create one if one doesn't exist. Once the room is full it will start the match.
Have an AI button that allows you to go to a lobby that you can play against AI.
Moved the GameModes to a folder and created a new Scriptable object for Card preparation.