Game Play Update
I altered the Card class to not be a scriptable object but a component that I can change and throw onto a card prefab.
I changed a lot of the PhotonGameManager script to begin to handle all the functionality it is going to need to to start playing through the game.
All my changes are in teh 1VOnline scene.
I created a CardDataBase that I am working through to having all the card information in the game.
I created a playerbase script that will act as a hub that maintains it's own health and updates itself based on damage taken.
I started to create a UIButtonHold script that has a little bit of the Touch code that can be used when getting everything to work on phones.
I started creating a UIConfirmation script that will get a confirmation passed and will wait for the confirmation or cancel to fulfill the command and will update the UI according to what it is being asked to confirm.