Match Update
I removed the generic Card prefab and created 3 separate ones to handle the card sizing problem.
DraftCard, HandCard, ZoomCard
I created a Zoom Card area that passes in a card to populate it with.
Plenty of changes in the 1VOnline scene.
Updated the CardDataBase to better utilize the cards and is my sole PhotonView object that expresses what needs to happen to all the cards between clients.
Updated the PhotonGameManager which handles all state issues.
Added a CardFunction script that calls out when a card has been "Selected"
Converted the Card script to a data only type script and created a CardData script to be the physical representation on the cards.
Currently have it where, the Network master picks who goes first.
It then tells the other(s) who they are.
Also got it to populate the hero draft, and give it over to other's to populate it evenly.