PvP Match Update
I adjusted the CardZoom prefab to use a Deselect function now on CardFunction.
This removes the zooming bool so it can be reapplied for allowing a card to actually be selected.
Removed all Actions between CardDataBase and PhotonGameManager.
I replaced them with direct method calls.
I adjusted both CardDataBase and PhotonGameManager to handle turn switches and drafting the first two heros, and adding a random hero to the last players hand.
Once the Hero draft phase is over it switches to the Ability Draft phase and then populates the draft area with the abilities.