HEROSelection Update
I began to build the visuals behind the HEROSelection options.
I created the Conformation UI that will pass what the asked request is, and if confirmed, then will execute the desired result.
I have Started building the framework for how each element will be executed.
I added a CardDeck enum for the CardDataBase to make it easier to get the desired deck.