Core Loop Update
Worked on building the Enhance function because it would get me to draw cards, and play cards from the hand.
I got the card collect amount ui set up.
I got the card drawing to the hand set up.
I set up a slider that will create an "offset" to rotate through the cards and set the active card to the most forward card.
Prepaired for  when a card will be played.