Enhance Update
I added a panel to the back of the card zoom so that the options would stick out.
I added Ui elements as needed to work with the Draw and play functions.
I added an auto draft turn that is currrently turned on permenantly to not have to draft all the cards manually while testing.
I added a lot of Debug.Logs to the code to see what was happening and what wasn't getting executed.
I added a turn indicator text to give a visual que who's turn it is if I wasn't paying attention.