Core Update
Added all the Hero Images to the image file, and put them in their respective image place.
Created a prefab for the heros when they are on your side and when they are one the opponents side.
Did a lot of work in the game UI to get everything to fit right.
CardData now listens for a targeting call, and highlights it's physical card(int theory).
I added an ability counter array and a target button for the CardData script.
CB knows each hero field. It also has the appropriate prefabs to instantiate one for that field.
It calls out a Target Action that GM and CardData listen for.
Set Up a spawning of the Hero cards, though I don't think it is actually hooked up yet.
Figured out how to dynamically call the value of the slider by changing the argument type to  System.Single but it comes in as a float.
Was working out a switch case based on all the ways to "Play A Card".
Added a Wait Enum to GamePhase, because I want the player to not do as much if they are waiting for their time to respond.
Needed an extra Hand-specific-zoom to be able to rotate the cards in the hand while zoomed on them.
Added a TurnCounter that will track how many more "interactions" or card plays the player still has remaining.