Drafting Update
I added an AI and AutoAbility Draft to the debug window.
I explicitly declared the text variables in the CardData to only change if they are not null.
I added AiDraft to the CB, to allow for the abilities to be drafted in the rule manual order.
To do this I added two list of the cards to predefine the cards drafted to ensure the balance.
I added the ability to differentiate the differing players by an Enum.
I also added a Phase Indicator to allow the player to see what phase they are in.