Enhancement Update
I added a List of Components that the Card Data will hold for all of it's abilities.
You can attach abilities and enhancements to either side, and it will show up for the other player.
Abilities will be expressed by a golden circle above the Hero that has it.
Enhancements are removed from the hand when they are used.
The opponents hand count is shown at the top right of the screen.