Hero Update
The decision board no longer shows up for both when the ability draft ends.
Send over the reserve picks when drawing from the reserve deck.
Sends Ghost array elements when the HQ needs less than 2 elements when filling.
Added a turn gauge to keep track of how many turns have passed.
Deactivating the reserve button when the reserves are depleted.
Added a few more regions to better separate what each grouping of methods are for.
In last test, I got an out of range exception when trying to play a hero to the field, it also didn't remove it from the hand and didn't lower the turn counter, need to look at the buttons for Enhancement again.