Recruit && Enhance Update
The Recruit can grab two heros, one from either HQ or reserve or both from one.
The Reserve button becomes unusable when there are no more heros.
The HQ repopulates as it should regardless how many were in it or in reserves.
Enhance no longer switched to wait then HERO after the turn ends for playing all types of cards.
When a card is played it is removed visually from the hand, as not to accidentally leave the copy behind when looking back at the hand.