Overcome Update
Added an Overcome Target button for the field hero prefabs.
Added an Overcome panel with a switch from attacking to defending characters,
a calculate battle button and end turn button.
CardData now registers a placement.
It looks for an activation/deactivation of Overcome.
It switches functionality when it switches from att/def.
It handles damage checking from received damage.
CB has a function that checks whether a card is yours or not.
CardFunction's OnHeroSelected action sends a CardData instead of Card.
GM holds more UI elements for the Overcome state.
It sends when Overcome is turned on and switched.
When coming out of it, it should turn all Overcome UI off.
It gained the Calculate Battle and SwitchAttDef methods.
It responds to hero selection and differentiates based on if it is your card and what phase of Overcome you are in.
I removed card zooming while in Overcome.