Card Update
I added all the Card Images and began to integrate them into the game.
It took some reworking of the Cards and CardData.
I needed to rework the Prefabs where all the cards show up, the field, hand, hq etc.
Also had to start undoing the hero selection in all of the different places.
I needed to fix the hero on field number issue.
Still working on the targeting. Will need to expand it to work for hero selection.