Hero Selection Update
I reinstated the various Overcome Calls.
I put a lot of color changes throughout CardData to see what a card is doing "state" wise.
I made a few adjustments to the 1VOnline scene to better see the overcome option board.
I added a bool check on the Targeting to turn off targeting.
Better built the AttDef system within GM.
(Might want to add a color change to a card that is targeted in Overcome)
Updated the UIConfirmation to use the bool check on targeting.