Overcome & Heal Update
Updated the CardData to handle various scenarios of a hero taking damage.
Updated the CB to check card states.
Updated the GM to handle all interaction with the Heros during the Overcome and Heal states.
Heros can be destroyed, block, exhausted through attacking or being attacked for over half, and can be healed.
May need to double check defense is halfed when exhausted.