HEROSelect update
I added a check for when the turn starts for each of the HERO selections based on whether or not they can be used.
Added a couple things in for the win condition.
Removed all HeroSelect methods since it was left behind.
Updated the exhaust and destroy of heros.
Added all the new pieces into the scene.