Ability Update
Started building more of the Ability functionality.
All the functionality of Absorb should work, needs tested.
Due to my decision to go with inheritance for abilities I had to adjust all Ability scripts to now inherit from Ability.
Made a UI section for the abilities of a character when zoomed. (Not tested)
Need to hook up all changes to the cloud.
Created a Prefab for the abilities so they can be added and removed upon zooming.
Created two new scripts that will handle the ability information for the player's viewing. UI elements.