Ability Update
Added all the HeroSpecific Abilities by name.
Set up the server functions from last addition of abilities.
When UICharacterAbility is created it is given an ability to take after.
Adjusted GM's UI for getting and removing the UI ability tags.
CB had to redefine the Ability spawn since the ability is a monobehavior and not a struct.
Added specific abilities and a list of them to the CArdData and methods to adjust them