Feat Update
I set all the code necessary for each of the four feats to be usable.
They need to be tested.
Also adjusted the characterabilityvisual prefab to no longer have two calls on the description button.
Updated the two ui ability components to better use the single call.
This needs tested to ensure that the description comes up and it is correct.
Did a bit of adding to the CB and GM for ability flexability.