Ability Update
I went through all hero and normal abilities and added their name and the ability description as well as type.
Added a few Debug.Logs into the Feat cards to better see where they are working or not.
Changed the UIAbilityDescriptor script location in scene and gave it a reference to the descriptor panel.
Adjusted  the way cards are removed from the visual hand, hopefully fixed the random card deletion.
Added a null catch in CardData when clearing abilities and enhancements if there weren't any.
May need to test enhancements again for that.