Ability Update
Updated the Card Description prefab, to set the description inactive on an X button press.
Adjusted those changes in the scene.
Added debug calls in the generic Ability methods.
Added the base call to each of the feat cards.
I added a null check on the enhancement list in CardData, that will create a new one if it is null.
CB the reveal hand call wasn't delegated right on awake.
CB commented out a RemoveCardFromHand.
CB added debugs to RemoveCardFromHand to find why visual hand count is off.
Added Ability name debug in setting active ability.