Ability Update
Adjusted all the current working ability scripts to have a virtual and override Awake function so that all listeners are inherited.
Put a character set bool in CardData so that it only is allowed to call it once instead of ever override.
Also set the exhausted bool to true before destroying.
In GM set opponent exhausted to true if the reference is ever found null, having been destroyed.