Ability Update
Added a few new Passivetypes to the enum in Ability.
Added a OncePerTurn and CanActivate bool in Ability.
Added OnDiscardCard, OnActivateable, OnHoldTurn and OnPreventAbilitiesToFieldForTurn in Ability.
Added a resetOncePerTurn listener in Ability and the respective method to fire.
Corrected Bolous single diget defense by adding * 10.
Started working out the code for Eng, Gambito, Grit, Hindra and Ignacia.
Added Sudo-code to Isaac, Izumi, Kay, and Kyauta.
Added the Protected override to Pay the Cost.
CardData got two new enhancement modifiers.
CardData has a method that sets the card values based on all the modifiers.
CB has a static herosFatigued counter that is updated when a hero is or isn't fatigued.
CB listens for HandleCardForceDiscard, and should deal with it appropriately.
CB hand check increased to < 8 instead of 7.
CB has the code to silenceAbilityToField.
GM got bEndTurn, canPlayAbiityToField bools and abilityPlaySilenceTurnTimer counter.
GM sends out an action to reset abilities that are one time use.
GM listens for OnActivateable, OnHoldTurn, and OnPreventAbilities...
Everywhere that a turn would just end now has a 3 second delay that will extend if someone is waiting to finish a move.
GM allows for an ability to be able to target "theoretically" in any phase.
GM timers count down when a turn is switched from control.