Abilities Update
Added a button to the CharacterAbilityVisual Prefab to activate activatable abilities.
Made some changes to the scene.
Added a passiveCheckable bool to Ability and used it in awake to only set up abilities if it is a card on the field.
Added ActivateAbility method to Ability to call it for the activate abilities.
Added passiveCheck to Akio, Ayumi,  Bolous, Christoph, Eng, Gambito, and Grit.
May actually need to remove those due to the check in Ability.
Eng's ability can only be triggered if the amount of exhausted people is higher than 0.
Added a Debug to Grit.
Added the first ActivateAbility call to Hindra.
Upon a character destroyed, I removed the abilities that were associated with the character from the Master Ability list in CB.
Added a Passive call for exhausted characters after the call in CB.
Changed GM bEndTurn to true instead of false as default.
Removed the EndTurnDelay at the beginning of GM.
Started organizing GM with regions.
Added a check to jump from passive calls if we are in a draft phase.
Adjusted the delay time before turn over to 1f instead of 3f.
UICharacterAbility listens for ability resets to turn on and off the activate button.
UICharacterAbility captures a new button to activate activatable abilities.