Ability Update
Added a new action in Ability to turn off the opponents turn wait.
Also put the adding the ability to master list in awake.
Removed all code in the Ability Passive Check.
Removed the PassiveCheckable check in Akio. Ayumi, Boulos, Eng and Christoph.
Added the new Turn off opp turn hold after christoph's ability executes.
Added a couple debug's to see where it is and isn't working.
Added the turn hold in the activation section of Christoph's ability.
christoph's ability works but not when he dies, need to create a pause on death until abilities are resolved.
Added the listener for the opp turn hold off in CB.
Added a function that turns it off.
Also adjusted the Ability Hand Over section a bit.
Added a PopUpUpdater in GM.
Added the turn hold stuff.