Code Refactor
Renamed the PhotonGameManager to Referee.
Renamed the PhotnGameManager prefab to Referee.
Due to the name change of the script all references in scene had to be reapplied.
Added a bool check to the Ability Action "OnHoldTurn" for checking whether it is my turn from where the ability is getting fired from.
Replaced all PhotonGameManager references for Referee in all scripts that addressed it.
Christoph got both referee reference fix and OnHoldTurn bool.
Eng also got the OnHoldTurn bool.
Changed a method name from RemoveEnhancements to RemoveAbilitiesAndEnhancemenst.
CardData got the Referee reference fix.
Did some major refactoring of CardDataBase to organize the code and remove any unnecessary calls to Photon and the new referee script, all code pulled was added to the new PhotonInGameManager script.
All CardDataBase calls to the PV are now to a reference to the new PhotonInGameManager.
The Referee was also diced up and removed all Photon calls and added a reference to the new PhotonInGameManager if necessary.
UICharacterAbility got the Referee reference fix.
UIConfirmation got the Referee reference fix.