Boss Meeting Update
I have moved elements around the scene as had been asked.
I added a turn counter back into the Referee script and check it in the DataBase for the enhance stage.
I added a new state called PreSelection where a player can look at the board before taking an action.
I also added a draw button for the player deck which tells how many cards can be drawn and how many cards are left in total.
I also added this function to the reserve deck as well when you need to draw from it during recruit.