Turn Sequence Update
I moved the End Turn Button from the pause menu to the screen at all times.
I also added a reference for it in Referee and turn the intractability on or off depending on if it is my turn or not.
Also finished adding all the places that PreSelection handled information and PostAction as well into Referee.