Game Phase Update
I removed PreSelection as a phase from CB and Ref.
Adjusted the HeroSelect game object to be a container that holds the interactives and a button which will turn the Interactives off or on, so that one can choose to look around the scene prior to selecting a hero action.