Card Function Update
Ref got an update so that a player whose turn it wasn't could not collect a card during the drafts. The gameobject was removed for some reason, and as such, I added it back in without knowing why it was removed in the first place.
After a player has selected one of the Hero Options it will simply hide the hero menu while confirming the decision.
While in Enhance, the player can collect cards until they play a card, after which the button will be deactivated.
Struggling to get the uI to scale right into the build.
Adjusted the build settings so that the player window is set to a particular size and can be scaled.
Added the debut button to the pause menu so it wasn't always on screen.