End Turn Update
Trying to make the scene more pleasing by getting rid of the end turn button if it isn't your turn, Code seems to be simple and clean, but it doesn't ever seem to turn it on when it's been turned off.
Moved some anchor points.
Moved the phase and turn indicators to the bottom right since the top was getting crowded.