Hand Update
Added to the Hand button to populate the draft area with the hand cards.
Start without all the usual visual buttons, like the hand, reserve, and deck info.
Added the turn count for allowing the player to play up to the count amount for a turn.
Fixed the issue with the end turn button not showing up.