Game Loop Update
I needed to reset everything in my brain because it has gotten too big to go back at it trying to piece it all together.
I made a new script that I have called CodeDumpRefactorZone and have started by dumping all the code that was in Referee to walk through the code getting each begining piece to work on it's own then adding in the other features as needed making sure it all works.
I have since been able to Reset the code so that I can Declare a player, turns, and allow the drafts to be displayed and interacted with.
However, the ability draft doesn't switch turns when a card is collected... not sure how I used to switch turns but I will add that in next time.
Future: Fix the ability draft, start making sure enhance works. Need to incorporate ability flags throughout play..