Action Response Update
CB I added a call to Photon that shows a card and awaits a response.
Also calls GM for waiting for a response and locks the player from interacting with things.
Added a function that gets called from Photon in CB to show the opponents card that was being played.
Pulled a couple methods out of the RefactorZone and put them back into GM
In Photon I added a listener to the Response timer to shut it down if the player didn't respond fast enough
Also added a call for when a response was actually made
Currently this call only passes on a bool value so I may need to expand
GM got a reference to the Response panel I made that has two buttons, yes or no.
May need to make sure that if the timer runs out the visuals should not be able to continue
Added the turn counter decrement method from CDRZ(refactor zone)
GM also got a method to finalize the step for showing opponents card
Added a couple bools into the UIResponseTimer to make sure calls aren't being called every cycle
Also added a stop timer, should we have a response before it ends
Also added a bool for starting the timer so that is should show on both screens just to let the waiting player know they are waiting on a response