Turn Switching Update
I changed some of the code in CB so that only player 1 will end the Ability draft if the autodraft is selected.
This also helped to eliminate the potential for player two to accidentally be on their turn but in the wait stage.
Also added a TurnOnGameElements method to be called for similar calls within the EndAbility method
In GM I Renamed EndTurn to EndTurnFromButton so it is more clear what is using the call.
I added several generic calls and a true switch turn which updates the opponent when the button is pressed.