Enhancement update
I added a check in the CardData for Opponent characters so that they don't add the abilities to their character cards.
In CB I added a CheckForHerosOnField method that returns a bool value so that we can know if we can play an enhancement or ability, if there isn't anything on the board, then they will not be allowed to play them.
In Ref, I added an extra args, for HandleCardButtons that allows the function to know what card we are working with and can access all card info for checks.
Also added the card null buttons if the card is an opponents card being shown.
In the HandleCardButtons method I added a check for card type and if there are viable targets on the field, or if the turn counter has run out, and if so, turn on or off the play button.