Overcome Update
I altered the HeroField and HeroFieldOpp prefabs to have a particle system that I am activating through Referee but controlled in CardData of the instance.
With this, I also removed the End Turn button that was on the Overcome UI since it is now always present.
CardData has a public method for setting the color and turning on the particles, and for stopping it.
I added all the old Referee Overcome methods and reworked them into the script so that they work now.
As such, I removed all the above code from the CodeDump script.
The overcome color changes need to be looked at because you can select things out of order and it messes up the colors. However, the overcome works as it is supposed to without the visual issues.