Slight Code Refaactoring
CardDataBase had a few things that were more modular and were used outside of the draft but the elements all still referred to draft elements. This has been changed for a universal terminology like CardDisplayArea.
In Scene, I turned off the Discard button for now, since it isn't needed at the beginning of the game during draft, though I haven't done anything with it yet.