Aisaac Update
Aisaac's ability now functions as it is supposed to, only triggering on the player who controls him, and can only grab cards that have been there prior to the combat that lead to the ability trigger.
This comes with a bug, you can accidentialy copy heros, if the bool isn't turned off right. Will need to further look into this.