Mace Update
Got the Mace ability working, added a call to the ability parent for when an action is complete and there is no communication.
Added the Mace passive stuff in the UIConfirmation and Referee.
Also adjusted the Isaac and Ayumi abilities to use a more universal Action that takes a string with their respective requests.