Yasmine Update
Yasmine now can play a card to the field from your hand as long as she has just been the target of a battle.
Had to update the UIConfirmation and the CardDataBase in order to get her to work.
Fixed a bug that came from the previous update that would build the HQ and pull from the reserve when doing the draft.