Zoe Update
Zoe now works as she is supposed to and can only be activated by the controlling player during the HEROSelect phase.
Updated the UIToggle so that it would force update the HEROSelection since, you could heal everyone before you select and the Overcome phase would be locked out, and the Heal phase would still be active.
Ability pays attention to placement and can hand it off if it needs to.