Accelerate Update
I had to make a load of changes in order to get the Accelerate ability to function as it is supposed to.
With that, we now have a working discard system, at least the discarding it from the hand.
Added the discard button for discarding.
Also added a scrolling animation to the card select so that I can actually go find any card.