Backfire Update
Another big update with canvas changes so that we can have a secondary ability list location to display and to activate them. I think it actually uses the characters ability to activate, but that should be fine since that is the only time it happens?
At least tested it on a character that got exhausted and then coupled with opposing Eng's heal ability which worked.