Boost Update
I built a working system to allow for player selection.
Had to add a few things to the scene to make it work.
Boost now only works when you actually have 2 or more cards.
CardDataBase does everything it needs to, except give the opponent cards right now.
UIAbilityDescriptor turns the button for activation on and off when it needs to.
UIConfirmation works with Boost.