Collateral Damage Update
Updated the Collateral Damage effect where after combat with a character who has it attached that did the damage and has defeated someone, you can chose to exhaust another hero as well.
Updated the Names in CardBase so they are consistent.
Made a change to targeting with ref, so you can target with the ability during combat.
UIConfirmation needed a call from the ability when triggered.