Convert Update
Added a bool to Collateral Damage so that I can ensure it is the correct ability being activated.
Added the same kind to Convert since it also needed unique activation.
Convert works as intended, though it does so without warning, may want a pop-up.
Referee, Photon and DDataBase all got the changes so that you can convert/steal a character -minus the abilities.