Counter Measure Update
Counter Measure now functions as expected.
Card Data needed a momentary adjustment to give a defense boost that is temporary then will go away.
There was several versions of Counter-Measures that needed to all be the same so they were changed in the script, DataBase and Ref.
I built the script to couple with the already combat response system we made for Origin.
We leveraged the UIConfirmation system to make sure everything was the way it needed to be.